refactor rule directory
