+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/stream/StreamAggregateImplementationRule.java
@@ -1,6 +1,6 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
 
-package com.starrocks.sql.optimizer.rule.implementation;
+package com.starrocks.sql.optimizer.rule.implementation.stream;
 
 import com.google.common.collect.Lists;
 import com.starrocks.sql.optimizer.OptExpression;
@@ -15,10 +15,10 @@ import java.util.List;
 
 public class StreamAggregateImplementationRule extends StreamImplementationRule {
 
-    private static final StreamAggregateImplementationRule INSTANCE =
-            new StreamAggregateImplementationRule(RuleType.IMP_STREAM_AGG);
+    private static final com.starrocks.sql.optimizer.rule.implementation.stream.StreamAggregateImplementationRule INSTANCE =
+            new com.starrocks.sql.optimizer.rule.implementation.stream.StreamAggregateImplementationRule(RuleType.IMP_STREAM_AGG);
 
-    public static StreamAggregateImplementationRule getInstance() {
+    public static com.starrocks.sql.optimizer.rule.implementation.stream.StreamAggregateImplementationRule getInstance() {
         return INSTANCE;
     }
