+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/stream/StreamJoinImplementationRule.java
@@ -1,6 +1,6 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
 
-package com.starrocks.sql.optimizer.rule.implementation;
+package com.starrocks.sql.optimizer.rule.implementation.stream;
 
 import com.google.common.collect.Lists;
 import com.starrocks.sql.optimizer.OptExpression;
